EmployeeService implemented.

+++ src/main/java/com/kht/backend/service/impl/OrganizationServiceImpl.java
@@ -12,6 +12,7 @@ import com.kht.backend.service.OrganizationService;
 import com.kht.backend.service.model.OrganizationModel;
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.LinkedHashMap;
@@ -19,54 +20,61 @@ import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
 
+@Service
 public class OrganizationServiceImpl implements OrganizationService {
     @Autowired
     private OrganizationDOMapper organizationDOMapper;
 
     @Override
-    @Transactional
-    public Result increaseOrganization(OrganizationModel organizationModel) {
-        OrganizationDO organizationDO = new OrganizationDO();
-        BeanUtils.copyProperties(organizationModel,organizationDO);
-        int affectRow=organizationDOMapper.insertSelective(organizationDO);
-        if(affectRow<=0)
-            throw  new ServiceException(ErrorCode.SERVER_EXCEPTION,"添加机构失败");
-        return Result.OK("添加机构成功").build();
+    public Result increaseOrganization(OrganizationDO organizationDO) {
+        if(organizationDO == null){
+            throw new ServiceException(ErrorCode.SERVER_EXCEPTION,"机构信息错误");
+        }
+        int affectRow = organizationDOMapper.insertSelective(organizationDO);
+        if(affectRow <= 0){
+            throw new ServiceException(ErrorCode.SERVER_EXCEPTION,"新增机构失败");
+        }
+        return Result.OK("新增机构成功").build();
     }
 
     @Override
-    @Transactional
     public Result decreaseOrganization(String organizationId) {
-        int affectRow=organizationDOMapper.deleteByPrimaryKey(organizationId);
-        if(affectRow<=0)
-            throw  new ServiceException(ErrorCode.SERVER_EXCEPTION,"删除机构失败");
+        if (organizationId == null || organizationId.equals("")){
+            throw new ServiceException(ErrorCode.SERVER_EXCEPTION,"机构编号不存在");
+        }
+        OrganizationDO organizationDO = organizationDOMapper.selectByPrimaryKey(organizationId);
+        if(organizationDO == null){
+            throw new ServiceException(ErrorCode.SERVER_EXCEPTION,"机构不存在，不需要删除");
+        }
+        int affectRow = organizationDOMapper.deleteByPrimaryKey(organizationId);
+        if(affectRow <= 0){
+            throw new ServiceException(ErrorCode.SERVER_EXCEPTION,"删除机构失败");
+        }
         return Result.OK("删除机构成功").build();
+
     }
 
     @Override
-    public Result getOrganization(int pageNum) {
-        //待改
-        PageHelper.startPage(pageNum,10);
-        List<OrganizationDO> organizationDOList=organizationDOMapper.selectAll();
-        List<OrganizationModel> organizationModelList=organizationDOList.stream().map(organizationDO -> {
-            OrganizationModel organizationModel= new OrganizationModel();
-            BeanUtils.copyProperties(organizationDO,organizationModel);
-            return organizationModel;
-        }).collect(Collectors.toList());
-        PageInfo<OrganizationDO> page = new PageInfo<>(organizationDOList);
-        Map<String,Object> resultData = new LinkedHashMap<>();
-        resultData.put("organization_num",page.getTotal());
-        resultData.put("organizations",organizationModelList);
-        return Result.OK(resultData).build();
+    public Result getOrganizationList(int pageNum) {
+
     }
 
     @Override
-    public Result updateOrganization(OrganizationModel organizationModel) {
-        OrganizationDO organizationDO = new OrganizationDO();
-        BeanUtils.copyProperties(organizationModel,organizationDO);
-        int affectRow=organizationDOMapper.updateByPrimaryKeySelective(organizationDO);
-        if(affectRow<=0)
-            throw  new ServiceException(ErrorCode.SERVER_EXCEPTION,"更新机构失败");
-        return Result.OK("更新机构成功").build();
+    public Result modifyOrganizationInfo(OrganizationDO organizationDO) {
+        OrganizationDO organizationDO1 = organizationDOMapper.selectByPrimaryKey(organizationDO.getOrgCode());
+        if(organizationDO1 == null){
+            throw new ServiceException(ErrorCode.SERVER_EXCEPTION,"机构不存在，不能修改信息");
+        }
+        int affectRow = organizationDOMapper.updateByPrimaryKeySelective(organizationDO);
+        if(affectRow <= 0){
+            throw new ServiceException(ErrorCode.SERVER_EXCEPTION,"机构修改失败");
+        }
+        return Result.OK("修改机构信息成功").build();
     }
+
+    @Override
+    public Result getOrganizationUser(int pageNum) {
+        r
+    }
+
 }
